+++ src/repl_tooling/repl_client/clojure.cljs
@@ -258,10 +258,12 @@
                    (js/setTimeout #(resolve cljs-repl) 500)))))
 
 (defn disable-limits! [aux]
-  (.then (eval/eval aux ":disable-limits")
-         #(eval/eval aux
-                     (unrepl-cmd (-> aux :session deref :state)
-                                 :print-limits
-                                 {:unrepl.print/string-length 9223372036854775807
-                                  :unrepl.print/coll-length 9223372036854775807
-                                  :unrepl.print/nesting-depth 9223372036854775807}))))
+  (if (instance? Evaluator aux)
+    (.then (eval/eval aux ":disable-limits")
+           #(eval/eval aux
+                       (unrepl-cmd (-> aux :session deref :state)
+                                   :print-limits
+                                   {:unrepl.print/string-length 9223372036854775807
+                                    :unrepl.print/coll-length 9223372036854775807
+                                    :unrepl.print/nesting-depth 9223372036854775807})))
+    (.resolve js/Promise {:result nil :as-text "nil"})))