+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.1
 - Fixed https://github.com/mauricioszabo/atom-chlorine/issues/150
+- Alpha support for nREPL
 
 ## 0.4.0
 - Add support for more than one form eval on Clojure REPL